BlTest

+++ BL/BO/Order.cs
@@ -22,7 +22,7 @@ namespace BO
         public DateTime? ShipDate { get; set; }
         public DateTime? DeliveryDate { get; set; }
         public List<OrderItem?>? Items { get; set; }
-        public double TotalPrice { get; set; }
+        public int TotalPrice { get; set; }
 
         //public override string ToString() => $@"
         //Order Id={OrderID}
